split debugging function
