+++ cosmos.decorator.tsx
@@ -1,0 +1,8 @@
+export default ({ children }: any) => {
+  return (
+    <>
+      <script src="https://cdn.tailwindcss.com" />
+      {children}
+    </>
+  )
+}
