+++ lib/index.ts
@@ -1,4 +1,7 @@
 export * from "./types"
+export * from "./graph-network-transformer/GraphNetworkTransformer"
+export * from "./operations/configureOperationCostFn"
+export * from "./operations/operation-types"
 export * from "./debug/getGraphicsForBpcGraph"
 export * from "./graph-utils/getGraphBounds"
 export * from "./graph-utils/getPinPosition"
