+++ tsconfig.json
@@ -9,7 +9,8 @@
     "allowJs": true,
 
     "paths": {
-      "lib/*": ["./lib/*"]
+      "lib/*": ["./lib/*"],
+      "lib": ["./lib"]
     },
 
     // Bundler mode
